deleted unnecesary controller
