+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/routing/comment.yml
@@ -1,6 +1,6 @@
 pim_comment_comment_delete:
     path: /comment/{id}/remove
-    defaults: { _controller: pim_comment.controller.comment:deleteAction }
+    defaults: { _controller: pim_enrich.controller.rest.product_comment:deleteAction }
     requirements:
         id: \d+
     methods: [DELETE]
